requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Cors>=4.0.0,<5.0.0
Flask-JWT-Extended>=4.6.0,<5.0.0
Flask-Mail>=0.10.0,<1.0.0
flasgger>=0.9.7.1,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=1.10.0,<2.0.0
email-validator>=2.2.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
Werkzeug>=3.0.0,<4.0.0